Move common slab memory implementation into src/chip8
Move font into separate header to make it usable with other renderers.
